fix: move assertValidPortMapping to lib/validation to fix coverage threshold
Importing from tools/run.ts pulled the full tool handler into coverage
scope, causing branch coverage to drop below 70%. Moving the validation
function to its own file isolates it.
